# docs/swagger.yaml
basePath: /
definitions:
  error.errorResponse:
    properties:
      message:
        type: string
    type: object
  error.errorResponseData:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  model.ResponseSuccess:
    properties:
      data: {}
      message:
        type: string
    type: object
...
host: https://mediumuz.herokuapp.com
info:
  contact:
    email: phapp0224mb@gmail.com
    name: Bakhodir Yashin Mansur
  description: API Server for MediumuZ Application
  termsOfService: https://github.com/BakhodiribnYashinibnMansur/Medium_uz
  title: MediumuZ API
  version: "1.0"
paths:
...
  /api/ghost/search:
    get:
      consumes:
      - application/json
      description: Search post by search text
      operationId: search-universal
      parameters:
      - description: 'Offset '
        in: query
        name: offset
        type: integer
      - description: 'Limit '
        in: query
        name: limit
        type: integer
      - description: search text
        in: query
        name: search
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      summary: Search  Post By search text
      tags:
      - Search
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"